+++ kikuchipy/pattern/correlate.py
@@ -25,14 +25,14 @@ from typing import Union
 from dask.array import Array
 import numpy as np
 
-# from kikuchipy._util import deprecated
+from kikuchipy._util import deprecated
 
 
-# @deprecated(
-#    since="0.4",
-#    alternative="kikuchipy.indexing.similarity_metrics.ncc",
-#    removal="0.5",
-# )
+@deprecated(
+    since="0.4",
+    alternative="kikuchipy.indexing.similarity_metrics.ncc",
+    removal="0.5",
+)
 def normalized_correlation_coefficient(
     pattern: Union[np.ndarray, Array],
     template: Union[np.ndarray, Array],